Cleanup before updating main to develop (#27)
* Moved DragonConfig to its own file

* Consolidated text utils into one file

* Added Source to SpellHeader

* Added more function documentation + 1 spell test

* Updated Header, Drawer, and Footer

* Removed "Return to Home" from pages

* Added more padding at bottom of layout
