+++ src/lib/dragon/dragon-stats.ts
@@ -1,5 +1,4 @@
 import {
-	DragonConfig,
 	COLOR_TO_ALIGNMENT,
 	AGE_TO_SIZE,
 	SIZE_TO_HIT_DIE,
@@ -14,11 +13,11 @@ import {
 	abilityMin,
 	abilityMax,
 	scoreToMod,
-	numberWithSign,
-	expectedDiceResult,
-	capitalizeFirstLetter
+	expectedDiceResult
 } from '.';
+import { capitalizeFirstLetter, numberWithSign } from '$lib/text-utils';
 import type { Age, Color, RGB, Size, Die, ProficiencyLevel, PronounsConfig } from '.';
+import { DragonConfig } from './dragon-config';
 import { DRAGON_VALS, type DragonVals } from './dragon-vals';
 import { type CR, CRNumberToString, CR_TABLE } from './challenge-rating';
 import { type SpellLevel, SPELL_LEVELS } from '$lib/spells';
