+++ src/lib/index.ts
@@ -1,20 +1,4 @@
-import { normalizeString } from './dragon';
-
-// https://stackoverflow.com/a/13627586/17934762
-export function numberWithOrdinalSuffix(i: number): string {
-	const j = Math.abs(i) % 10,
-		k = Math.abs(i) % 100;
-	if (j == 1 && k != 11) {
-		return `${i}st`;
-	}
-	if (j == 2 && k != 12) {
-		return `${i}nd`;
-	}
-	if (j == 3 && k != 13) {
-		return `${i}rd`;
-	}
-	return `${i}th`;
-}
+import { normalizeString } from './text-utils';
 
 export const NAV_LINKS = [
 	{
@@ -28,17 +12,18 @@ export const NAV_LINKS = [
 		children: []
 	},
 	{
-		href: '/spells/',
-		text: 'Spells',
-		children: []
-	},
-	{
 		href: '/about/',
 		text: 'About',
 		children: []
 	}
 ] as const;
 
+/**
+ * Returns a unique ID string for the given Nav Link. Uniqueness is guaranteed by testing.
+ * @export
+ * @param {(typeof NAV_LINKS)[number]} link
+ * @return {*}  {string}
+ */
 export function getNavLinkId(link: (typeof NAV_LINKS)[number]): string {
 	return `navLink__${normalizeString(link.text)}`;
 }
