+++ src/routes/prismatic-dragons/+page.svelte
@@ -10,4 +10,3 @@
 
 <h1>The Prismatic Dragons</h1>
 <p>This page will contain info about the Prismatic Dragons.</p>
-<p><a href="/">Return to Home.</a></p>
